Add Task.java which contains Task class
This is more user-friendly as it allows user to work at the level of
objects while ignoring the data attributes and method implementation
details. User can store multiple attributes within a Task object and
the attributes are only accessible using the Task object's interface.

Let's create class Task and construct attributes getters and setters
and other relevant methods for the class Task.

This replaces the usage of the String to represent the task and the
Task list (array) is used instead of a String array and allows user to
interact the objects via different methods.

Replace method processForThreeSeconds() to processForOneSecond()

It is taking slightly long before the user can enter the next command.
Some user may be impatient and accidentally keys in command while the
chatbot Yq is processing an action. This could trigger error and bring
inconvenience to the user.

Reduce 3-seconds delay to a 1-second delay.

This reduces the waiting time of the user while giving more realism as
program requires time to process an action or command input by the user.

Create methods printList(), addTask(), markTask() and unmarkTask()

By compiling the codes into different methods, it is more readable,
cleaner and user-friendly as requested to display task list contents,
add task into the task list, mark task as done and mark task as not
done. Try and catch exceptions are added in markTask() and unmarkTask()
to prevent invalid input number by the user as it involves the user's
input to extract the Task with their corresponding position from the
Task list.

Make each action or command by the user as a method.

More methods may be implemented in the future and certain methods would
require some sort of output such as an Array (Task List). It is also
more convenient to add on more methods if more commands are needed for
the chatbot Yq.