Refine markTask() and unmarkTask() methods in Yq.java
The condition, whereby the task list is empty, located outside the
markTask() and unmarkTask() methods is unnecessary as it is already
accounted for in both methods. Also, it is inconvenient for the user to
keep entering the 'mark' and 'unmark' command whenever the user keys in
an invalid number to extract the task to be marked as done or not done.

Remove the exterior if-condition when the task list is empty and
implement a while loop which encapsulates the try-catch statements
within the markTask() and unmarkTask() methods.

A while loop encapsulating the try-catch statements in both methods
allows the user to input a valid number after keying in the wrong number
without the need to exit the markTask() or unmarkTask() method and key
in the command again.

Change methods from setIsDone() to setDone() and getIsDone() to isDone()
in Task.java

This is to improve readability and ensures that the method name follows
the coding standard.