Add child classes of Task class: ToDo, Deadline and Event
These classes are required for Level-4 increment and inheritance is used
as Todo, Deadline and Event classes have similar methods and attributes
as the parent Task class.

toString() method is included in all these classes to accommodate the
differences in the method implementation between the parent class and
the child classes via overriding.