Make some minor adjustments and refinements to ensure high code quality
This is to ensure that the codes adhere to the Java coding standards and
are of high code quality.

Add method definitions and explanations in Yq.java

The definitions and explanations of the methods allow reviewers and
users to understand at a greater detail in what the methods do and how
they work.