Make some minor code refactoring and other tweaks based on peer review
Some of the variable names are shortened for better readability.
Numbers are assigned with meaningful variable names to prevent magic
numbers.
To avoid arrowhead code style, isValidInputInteger() method, which
checks for valid input integer in both 'mark' and 'unmark' commands,
is being extracted from the isValidUnmarkCmd() and isValidMarkCmd()
methods, hence simplifying them.
The expected output in the EXPECTED.TXT are also adjusted based on the
changes made in Yq.java accordingly.