Modify the methods so they can handle errors by throwing exceptions
The methods returnValidCommand(), markTask(), unmarkTask(),
createTodo(), createDeadline(), createEvent() and printList() contain
exceptions which will be throw according to the error found in the
user's input.

Each of the method above has a corresponding method that contains
try-and-catch statements which handles the abnormality detected in the
user's input by printing the respective error messages after the
exception is thrown while an action is executed.