Refine methods that involve the task ArrayList
The parameter for the task ArrayList is removed from those methods and
the ArrayList variables present in the codes within those methods are
replaced with an inline ArrayList variable instead, which is the private
static final task ArrayList declared in the Yq class.

This ensures readability and the methods in Yq.class only function on
that specific task ArrayList.