Include methods and features related to deleting task
Methods such as deleteTask() and checkValidDeleteCmd() are added into
Yq.java. deleteTask() deletes the task based on the index selected by
the user and throws exception whenever an error is encountered in the
user input. Try-and-catch statements are added in checkValidDeleteCmd()
to ensure that the delete command is valid before the task is being
deleted from the task list.

Changes to other methods such as main() and checkValidCommand() in
Yq.java are made because they need to account for the 'delete' keyword.
The chatbot Yq has gained another ability to delete task from the
task list as requested by the user.